requirements.txt
streamlit>=1.35.0
streamlit-extras>=0.4.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
transformers>=4.41.0
torch>=2.3.0
pypdf2>=3.0.0
pymupdf>=1.24.0
python-docx>=1.1.0
pillow>=10.3.0
pytesseract>=0.3.10
opencv-python-headless>=4.9.0
graphviz>=0.20.3
matplotlib>=3.9.0
plotly>=5.22.0
python-dotenv>=1.0.1
accelerate>=0.30.0
ctransformers>=0.2.27